requirements.txt
pyparsing
networkx
pint
watchdog
pyramid
pyramid_jinja2
bs4
nose
reprep
PyContracts
decentlogs
markdown
#-e git+https://github.com/AndreaCensi/contracts.git@mcdp#egg=PyContracts
-e git+https://github.com/AndreaCensi/compmake.git@mcdp#egg=compmake
-e git+https://github.com/AndreaCensi/quickapp.git@mcdp#egg=quickapp
-e git+https://github.com/AndreaCensi/conf_tools.git@mcdp#egg=ConfTools
-e git+https://github.com/AndreaCensi/comptests.git@mcdp#egg=comptests
-e git+https://github.com/AndreaCensi/gvgen.git@master#egg=gvgen
# -e git+https://github.com/AndreaCensi/memos.git@master#egg=memos
#numpy
#matplotlib
